+++ DatadogCore/Tests/Datadog/Mocks/RUMFeatureMocks.swift
@@ -699,6 +699,12 @@ func mockNoOpSessionListener() -> RUM.SessionListener {
     return { _, _ in }
 }
 
+internal class FatalErrorContextNotifierMock: FatalErrorContextNotifying {
+    var sessionState: RUMSessionState?
+    var view: RUMViewEvent?
+    var globalAttributes: [String: Encodable] = [:]
+}
+
 extension RUMScopeDependencies {
     static func mockAny() -> RUMScopeDependencies {
         return mockWith()
@@ -713,11 +719,13 @@ extension RUMScopeDependencies {
         firstPartyHosts: FirstPartyHosts = .init([:]),
         eventBuilder: RUMEventBuilder = RUMEventBuilder(eventsMapper: .mockNoOp()),
         rumUUIDGenerator: RUMUUIDGenerator = DefaultRUMUUIDGenerator(),
+        backtraceReporter: BacktraceReporting = BacktraceReporterMock(backtrace: nil),
         ciTest: RUMCITest? = nil,
         syntheticsTest: RUMSyntheticsTest? = nil,
         vitalsReaders: VitalsReaders? = nil,
         onSessionStart: @escaping RUM.SessionListener = mockNoOpSessionListener(),
-        viewCache: ViewCache = ViewCache()
+        viewCache: ViewCache = ViewCache(),
+        fatalErrorContext: FatalErrorContextNotifying = FatalErrorContextNotifierMock()
     ) -> RUMScopeDependencies {
         return RUMScopeDependencies(
             featureScope: featureScope,
@@ -728,11 +736,13 @@ extension RUMScopeDependencies {
             firstPartyHosts: firstPartyHosts,
             eventBuilder: eventBuilder,
             rumUUIDGenerator: rumUUIDGenerator,
+            backtraceReporter: backtraceReporter,
             ciTest: ciTest,
             syntheticsTest: syntheticsTest,
             vitalsReaders: vitalsReaders,
             onSessionStart: onSessionStart,
-            viewCache: viewCache
+            viewCache: viewCache,
+            fatalErrorContext: fatalErrorContext
         )
     }
 
@@ -745,11 +755,13 @@ extension RUMScopeDependencies {
         firstPartyHosts: FirstPartyHosts? = nil,
         eventBuilder: RUMEventBuilder? = nil,
         rumUUIDGenerator: RUMUUIDGenerator? = nil,
+        backtraceReporter: BacktraceReporting? = nil,
         ciTest: RUMCITest? = nil,
         syntheticsTest: RUMSyntheticsTest? = nil,
         vitalsReaders: VitalsReaders? = nil,
         onSessionStart: RUM.SessionListener? = nil,
-        viewCache: ViewCache? = nil
+        viewCache: ViewCache? = nil,
+        fatalErrorContext: FatalErrorContextNotifying? = nil
     ) -> RUMScopeDependencies {
         return RUMScopeDependencies(
             featureScope: self.featureScope,
@@ -760,11 +772,13 @@ extension RUMScopeDependencies {
             firstPartyHosts: firstPartyHosts ?? self.firstPartyHosts,
             eventBuilder: eventBuilder ?? self.eventBuilder,
             rumUUIDGenerator: rumUUIDGenerator ?? self.rumUUIDGenerator,
+            backtraceReporter: backtraceReporter ?? self.backtraceReporter,
             ciTest: ciTest ?? self.ciTest,
             syntheticsTest: syntheticsTest ?? self.syntheticsTest,
             vitalsReaders: vitalsReaders ?? self.vitalsReaders,
             onSessionStart: onSessionStart ?? self.onSessionStart,
-            viewCache: viewCache ?? self.viewCache
+            viewCache: viewCache ?? self.viewCache,
+            fatalErrorContext: fatalErrorContext ?? self.fatalErrorContext
         )
     }
 }