+++ DatadogCore/Tests/DatadogObjc/ObjcAPITests/DDRUM+apiTests.m
@@ -134,5 +134,4 @@
 
 #pragma clang diagnostic pop
 
-
 @end