+++ DatadogRUM/Sources/Integrations/CrashReportReceiver.swift
@@ -23,6 +23,8 @@ internal struct CrashReportReceiver: FeatureMessageReceiver {
     }
 
     struct CrashContext: Decodable {
+        /// The Application launch date
+        let appLaunchDate: Date?
         /// Interval between device and server time.
         let serverTimeOffset: TimeInterval
         /// The name of the service that data is generated from.
@@ -41,6 +43,8 @@ internal struct CrashReportReceiver: FeatureMessageReceiver {
         var lastRUMViewEvent: RUMViewEvent?
         /// State of the last RUM session in crashed app process.
         var lastRUMSessionState: RUMSessionState?
+        /// The last global RUM attributes in crashed app process.
+        var lastRUMAttributes: GlobalRUMAttributes?
         /// The last _"Is app in foreground?"_ information from crashed app process.
         let lastIsAppInForeground: Bool
         /// Network information.
@@ -66,6 +70,8 @@ internal struct CrashReportReceiver: FeatureMessageReceiver {
         let realCrashDate: Date
         /// Current time, adjusted with NTP correction.
         let realDateNow: Date
+        /// Time between crash and application launch
+        let timeSinceAppStart: TimeInterval?
     }
 
     /// RUM feature scope.
@@ -127,14 +133,26 @@ internal struct CrashReportReceiver: FeatureMessageReceiver {
         let currentTimeCorrection = context.serverTimeOffset
 
         let crashDate = report.date ?? dateProvider.now
+        var timeSinceAppStart: TimeInterval? = nil
+        if let startDate = context.appLaunchDate {
+            timeSinceAppStart = crashDate.timeIntervalSince(startDate)
+        }
+
         let adjustedCrashTimings = AdjustedCrashTimings(
             crashDate: crashDate,
             realCrashDate: crashDate.addingTimeInterval(currentTimeCorrection),
-            realDateNow: dateProvider.now.addingTimeInterval(currentTimeCorrection)
+            realDateNow: dateProvider.now.addingTimeInterval(currentTimeCorrection),
+            timeSinceAppStart: timeSinceAppStart
         )
 
         // RUMM-2516 if a cross-platform crash was reported, do not send its native version
-        if let lastRUMViewEvent = context.lastRUMViewEvent {
+        if var lastRUMViewEvent = context.lastRUMViewEvent {
+            if let lastRUMAttributes = context.lastRUMAttributes {
+                // RUM-3588: If last RUM attributes are available, use them to replace view attributes as we know that
+                // global RUM attributes can be updated more often than attributes in `lastRUMView`.
+                // See https://github.com/DataDog/dd-sdk-ios/pull/1834 for more context.
+                lastRUMViewEvent.context?.contextInfo = lastRUMAttributes.attributes
+            }
             if lastRUMViewEvent.view.crash?.count ?? 0 < 1 {
                 sendCrashReportLinkedToLastViewInPreviousSession(
                     report,
@@ -165,13 +183,13 @@ internal struct CrashReportReceiver: FeatureMessageReceiver {
         using crashTimings: AdjustedCrashTimings
     ) {
         if crashTimings.realDateNow.timeIntervalSince(crashTimings.realCrashDate) < FatalErrorBuilder.Constants.viewEventAvailabilityThreshold {
-            send(crashReport: crashReport, to: lastRUMViewEvent, using: crashTimings.realCrashDate)
+            send(crashReport: crashReport, to: lastRUMViewEvent, using: crashTimings)
         } else {
             // We know it is too late for sending RUM view to previous RUM session as it is now stale on backend.
             // To avoid inconsistency, we only send the RUM error.
             DD.logger.debug("Sending crash as RUM error.")
             featureScope.eventWriteContext(bypassConsent: true) { context, writer in
-                let builder = createFatalErrorBuilder(context: context, crash: crashReport, crashDate: crashTimings.realCrashDate)
+                let builder = createFatalErrorBuilder(context: context, crash: crashReport, crashDate: crashTimings.realCrashDate, timeSinceAppStart: crashTimings.timeSinceAppStart)
                 let rumError = builder.createRUMError(with: lastRUMViewEvent)
 
                 if let mappedError = self.eventsMapper.map(event: rumError) {
@@ -231,7 +249,7 @@ internal struct CrashReportReceiver: FeatureMessageReceiver {
         }
 
         if let newRUMView = newRUMView {
-            send(crashReport: crashReport, to: newRUMView, using: crashTimings.realCrashDate)
+            send(crashReport: crashReport, to: newRUMView, using: crashTimings)
         }
     }
 
@@ -284,18 +302,18 @@ internal struct CrashReportReceiver: FeatureMessageReceiver {
         }
 
         if let newRUMView = newRUMView {
-            send(crashReport: crashReport, to: newRUMView, using: crashTimings.realCrashDate)
+            send(crashReport: crashReport, to: newRUMView, using: crashTimings)
         }
     }
 
     /// Sends given `CrashReport` by linking it to given `rumView` and updating view counts accordingly.
-    private func send(crashReport: DDCrashReport, to rumView: RUMViewEvent, using realCrashDate: Date) {
+    private func send(crashReport: DDCrashReport, to rumView: RUMViewEvent, using crashTimings: AdjustedCrashTimings) {
         DD.logger.debug("Updating RUM view with crash report.")
 
         // crash reporting is considering the user consent from previous session, if an event reached
         // the message bus it means that consent was granted and we can safely bypass current consent.
         featureScope.eventWriteContext(bypassConsent: true) { context, writer in
-            let builder = createFatalErrorBuilder(context: context, crash: crashReport, crashDate: realCrashDate)
+            let builder = createFatalErrorBuilder(context: context, crash: crashReport, crashDate: crashTimings.realCrashDate, timeSinceAppStart: crashTimings.timeSinceAppStart)
             let updatedRUMView = builder.updateRUMViewWithError(rumView)
             let rumError = builder.createRUMError(with: updatedRUMView)
 
@@ -313,7 +331,7 @@ internal struct CrashReportReceiver: FeatureMessageReceiver {
 
     // MARK: - Building RUM events
 
-    private func createFatalErrorBuilder(context: DatadogContext, crash: DDCrashReport, crashDate: Date) -> FatalErrorBuilder {
+    private func createFatalErrorBuilder(context: DatadogContext, crash: DDCrashReport, crashDate: Date, timeSinceAppStart: TimeInterval?) -> FatalErrorBuilder {
         return FatalErrorBuilder(
             context: context,
             error: .crash,
@@ -324,7 +342,8 @@ internal struct CrashReportReceiver: FeatureMessageReceiver {
             errorThreads: crash.threads.toRUMDataFormat,
             errorBinaryImages: crash.binaryImages.toRUMDataFormat,
             errorWasTruncated: crash.wasTruncated,
-            errorMeta: crash.meta.toRUMDataFormat
+            errorMeta: crash.meta.toRUMDataFormat,
+            timeSinceAppStart: timeSinceAppStart
         )
     }
 
@@ -366,7 +385,10 @@ internal struct CrashReportReceiver: FeatureMessageReceiver {
                 carrierInfo: context.carrierInfo
             ),
             container: nil,
-            context: nil,
+            // RUM-3588: We know that last RUM view is not available, so we're creating a new one. No matter that, try using last
+            // RUM attributes if available. There is a chance of having them as global RUM attributes can be updated more often than RUM view.
+            // See https://github.com/DataDog/dd-sdk-ios/pull/1834 for more context.
+            context: context.lastRUMAttributes.map { .init(contextInfo: $0.attributes) },
             date: startDate.timeIntervalSince1970.toInt64Milliseconds,
             device: .init(device: context.device, telemetry: featureScope.telemetry),
             display: nil,