+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Unreleased
 
+- [IMPROVEMENT] Crash errors now include up-to-date global RUM attributes. See [#1834][]
 - [FEATURE] `DatadogTrace` now supports OpenTelemetry. See [#1828][]
 - [FEATURE] Support for trace context injection configuration to allow selective injection. See [#1835][]
+- [FEATURE] `DatadogWebViewTracking` is now available for Obj-C. See [#1854][]
+- [FEATURE] RUM "stop session", "get session ID" and "evaluate feature flag" APIs are now available for Obj-C. See [#1853][]
 
 # 2.11.0 / 08-05-2024
 
@@ -660,9 +663,12 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1767]: https://github.com/DataDog/dd-sdk-ios/pull/1767
 [#1798]: https://github.com/DataDog/dd-sdk-ios/pull/1798
 [#1776]: https://github.com/DataDog/dd-sdk-ios/pull/1776
+[#1834]: https://github.com/DataDog/dd-sdk-ios/pull/1834
 [#1721]: https://github.com/DataDog/dd-sdk-ios/pull/1721
 [#1803]: https://github.com/DataDog/dd-sdk-ios/pull/1803
+[#1853]: https://github.com/DataDog/dd-sdk-ios/pull/1853
 [#1807]: https://github.com/DataDog/dd-sdk-ios/pull/1807
+[#1854]: https://github.com/DataDog/dd-sdk-ios/pull/1854
 [#1828]: https://github.com/DataDog/dd-sdk-ios/pull/1828
 [#1835]: https://github.com/DataDog/dd-sdk-ios/pull/1835
 [@00fa9a]: https://github.com/00FA9A